requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic

langchain-openai
langchain-community
lancedb
langchain-experimental

pandas